Create Full Master Data Group. Some Adjustment at MVC

+++ app/Database/Migrations/2023-08-04-041200_CreateBooksTable.php
@@ -10,7 +10,7 @@ class CreateBooksTable extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type' => 'bigint',
+                'type' => 'int',
                 'unsigned' => true, 
                 'auto_increment' => true
             ],
